==> R/script6.1_tratamentoMeta5.R <==

# Tentando substituir os códigos dos municipios pelo nome
municipiosSaeb = Meta5 |> dplyr::select(id_municipio) |>
  stringr::str_replace(id_municipio,"Bom Jesus das Selvas","Bom Jesus das Selvas")

table(municipiosSaeb$id_municipio) # Contando quantos municípios existem na base de dados

listaMunicipiosSaeb = c(6311031, 6311033, 6311034, 6311035, 6311036,
  6311038, 6311039, 6311040, 6311041, 6311042,
  6311043, 6311044, 6311045, 6311046, 6311047,
  6311049, 6311051, 6311052, 6311053, 6311057,
  6311060, 6311061, 6311065, 6311078, 6311079,
  6311081, 6311082, 6311083, 6311084, 6311087,
  6311088, 6311089, 6311090, 6311092, 6311093,
  6311094, 6311095, 6311096, 6311097, 6311099,
  6311100, 6311101, 6311102, 6311103, 6311104,
  6311107, 6311108, 6311110, 6311112, 6311113,
  6311115, 6311116, 6311117, 6311118, 6311119,
  6311121, 6311122, 6311123, 6311125, 6311129,
  6311130, 6311131, 6311132, 6311133, 6311134,
  6311136, 6311139, 6311141, 6311142, 6311143,
  6311144, 6311147, 6311148, 6311149, 6311151,
  6311152, 6311153, 6311154, 6311156, 6311159,
  6311160, 6311161, 6311164, 6311167, 6311168,
  6311169, 6311170, 6311171, 6311172, 6311173,
  6311174, 6311175, 6311176, 6311177, 6311179,
  6311180, 6311181, 6311188, 6311190, 6311198,
  6311199, 6311201, 6311202, 6311205, 6311207,
  6311209, 6311211, 6311212, 6311213, 6311214,
  6311216, 6311219, 6311225, 6311230, 6311231,
  6311233, 6311238, 6311242, 6311243, 6311246,
  6311248, 6311252, 6311255, 6311260, 6311262,
  6311270, 6311275, 6311276, 6311281, 6311282,
  6311286, 6311288, 6311289, 6311294, 6311301,
  6311302, 6311306, 6311307, 6311311, 6311312,
  6311314, 6311316, 6311317, 6311318, 6311324,
  6311325, 6311326, 6311327, 6311330, 6311331,
  6311332, 6311333, 6311334, 6311335, 6311336,
  6311337, 6311338, 6311339, 6311344, 6311345,
  6311346, 6311347, 6311348, 6311351, 6311352,
  6311354, 6311355, 6311356, 6311357, 6311367,
  6311369, 6311370, 6311372, 6311375, 6311376,
  6311378, 6311381, 6311385, 6311389, 6311390,
  6311393, 6311396, 6311397, 6311398, 6311402,
  6311406, 6311415, 6311416, 6311417, 6311420,
  6311423, 6311430, 6311431, 6311432, 6311434,
  6311435, 6311441, 6311444, 6311445, 6311450,
  6311451, 6311456, 6311457, 6311462, 6311466,
  6311470, 6311473, 6311475, 6311479, 6316601,
  6316604, 6316609, 6316611, 6316612, 6316614,
  6316615, 6316616, 6316621, 6316624, 6316631,
  6316635, 6316652, 6316654, 6316657, 6316659,
  6316660, 6316662, 6316663, 6316665, 6316666,
  6316667, 6316669, 6316670, 6316673, 6316674,
  6316675, 6316678, 6316682, 6316691, 6316693,
  6316699, 6316701, 6316709, 6316711, 6316723,
  6316724, 6316731, 6316737, 6316738, 6316740,
  6316741, 6316743, 6316744, 6316747, 6316750,
  6316759, 6316760, 6316769, 6316771, 6316772,
  6316779, 6316782, 6316783, 6316784, 6316786,
  6316800, 6316803, 6316845, 6316859, 6316895,
  6316897, 6316898, 6316902, 6316903, 6316904,
  6316905, 6316906, 6316908, 6316909, 6316918,
  6316922, 6316927, 6316932, 6316959, 6316963,
  6316967, 6316968, 6317001, 6317002, 6317007,
  6317014, 6317021, 6317027, 6317043, 6317045)

quantMunic = length(listaMunicipiosSaeb);quantMunic

# Não consegui encontrar os códigos correspondentes aos que estão presentes
# na base da Meta 5.
