==> R/script9_calcMeta8.R <==
# Meta 8: OFERECER EDUCAÇÃO EM TEMPO INTEGRAL EM, NO MÍNIMO,
# 50% (CINQUENTA POR CENTO) DAS ESCOLAS PÚBLICAS, DE FORMA A
# ATENDER, PELO MENOS, 25%(VINTE E CINCO POR CENTO) DOS (AS)
# ALUNOS (AS) DA EDUCAÇÃO BÁSICA.

# INDICADOR 8A: Escolaridade média da população de 18 a 29 anos de idade.

# INDICADOR 8B: Escolaridade média da população de 18 a 29 anos residente
# na área rural.

# INDICADOR 8C: Escolaridade média da população de 18 a 29 anos pertencente
# aos 25% mais pobres (renda domiciliar per capita)

# INDICADOR 8D: Razão entre a escolaridade média de negros e não negros na
# faixa etária de 18 a 29 anos
